requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.2.0,<6.0
django-cors-headers>=3.13,<4.0